Do not alter results of standard completion (if any)
When `bash-completion` package is in use and its helper functions
(_filedir*) succeed at completing the path, there is no need to invoke the
enhanced path completion.

This fixes the issue when the special characters in the path were escaped
twice resulting in invalid completion results.

Closes #5.

Thanks to StrangeAstronomer (https://www.reddit.com/user/StrangeAstronomer)
for reporting the bug.