Improve parsing of completion results
Bash completion provides both list of possible variants and a pre-selected
variant of the command line. The latter was lost in previous
implementation of complete() method.

Now complete() returns a custom object instead of string. This object can
easily be transformed into str() or set() for use in assertions. Data
members can also be accessed directly.